Ensure headers are installed even with the new multilib layout
The default multilib is now installed to a specific MULTISUBDIR
which thwarts the logic to only install headers for the top level
multilib. Change the detection logic so that either the MULTISUBDIR
is empty (which is the classic way of detecting) or check that the
MULTIDIRS is non-empty. Only one of these conditions will ever hold
true and will only be in the top level/default multilib build area.