Add a branch emulator.
Add a branch emulator that takes some instructions and context, and gives
the consequent program counter.

The original patch was written by Lisa Parratt. I perfored the code review
and the testing of MIPS32R2 and microMIPS32R3 under Qemu.

This needs more testing in order to verify that it works correctly for
big-endian machines, hard-float instructions and odd values of program
counter (Qemu ignores the last bit).