Wrap __excpt_ebase definition in PROVIDE to allow overriding
The ebase is supposed to support being overridden but was not
allowing it. This means an alternate user-controlled area can be
defined but the EBASE register will still be updated by the runtime
code. When overridden then there is a chance that an early exception
will go unhandled between setting ebase and entering application
code that sets up the associated memory.