nanoMIPS only changes
setjmp/longjmp
==============

The alignment of the jmpbuf is increased to 8 bytes and one additional
unused slot is available in the GPR register range.

memset/memcpy/strlen
====================
New C versions of these are used for nanoMIPS only. This is under
review.
The C version of strlen is also used for nanoMIPS. Also to review.

linker scripts
==============
Conditionally set the _gp symbol to be at the start of the small data
area for nanoMIPS.

syscalls
========
Force use of a 32-bit syscall for UHI operations.

micromips/mips16
================
Do not emit any nomips16 or nomicromips assembler directives as they are
unsupported.

ISA abstraction
===============
Use the correct form for JR.HB which no longer needs special delay slot
handling given there are no delay slots.
MSA is not usable with nanoMIPS yet.
Enable the TLB ASE for boot TLB code for nanoMIPS
Disable _mips_intpatch as it needs porting to nanoMIPS

multilibs
=========
Support the same controls as MIPS.

configury
=========
Add mt-nanomips - this just sets the -mno-gpopt and we use standard
fragments to choose size over speed