GTM19-493: UHI linker scripts fail to link with _isr_vec_count=1
This _isr_vec_count=1 is prevalent in MIPS Open FGPA examples.
Looks like forcing _isr_vec_007 in to the link was a mistake
and we always intended to only include the top-level _isr_vec
symbol.  This change has been reverted on multiple internal
branches, but someone still ended up carried forward on the
master.

libgloss/
	* mips/uhi32.ld: Retain _isr_vec instead of _isr_vec_007.
	* mips/uhi64_64.ld: Likewise.
	* mips/uhi64_n32.ld: Likewise.
	* mips/examples/romable_predef_xip/Makefile: Remove
	unnecessary definition of _isr_vec_007 to 0.