Cygwin: AF_UNIX: recvmsg: refactor
Move the code that peeks at a packet header into a new method
peek_pipe_poll.  This is like peek_pipe, but it polls until there's
data in the pipe, an error, or a signal.

Also move the code that copies data to the scatter-gather locations
out of the switch statement.

Both of these changes will facilitate a future commit to support the
MSG_PEEK flag.