Add build mechanism to share common header files between machines
So far the build mechanism in newlib only allowed to either define
machine-specific headers, or headers shared between all machines.
In some cases, architectures are sufficiently alike to share header
files between them, but not with other architectures.  A good example
is ix86 vs. x86_64, which share certain traits with each other, but
not with other architectures.

Introduce a new configure.host variable called "shared_machine_dir".
For ix86 and x86_64, define it to "shared_x86".  For a start, add the
common sys/fenv.h file to this directory.

Drop the Cygwin-specific fenv.cc and fenv.h file and use the equivalent
newlib functionality now, so we have at least one example of a user for
this new mechanism.  Drop _feinitialise from _dll_crt0 since newlib's
fenv initializes on the fly.

Signed-off-by: Corinna Vinschen <corinna@vinschen.de>